+++ market/src/main/java/org/mercatia/danp/jobs/LogicRefiner.java
@@ -20,9 +20,9 @@ public class LogicRefiner extends GenericJob {
 
 	@Override
 	public void simulate(Market market) {
-		float food = queryInventory("food");
-		float tools = queryInventory("tools");
-		float ore = queryInventory("ore");
+		double food = queryInventory("food");
+		double tools = queryInventory("tools");
+		double ore = queryInventory("ore");
 
 		boolean has_food = food >= 1;
 		boolean has_tools = tools >= 1;
