+++ market/src/main/java/org/mercatia/danp/jobs/LogicWoodcutter.java
@@ -19,8 +19,8 @@ public class LogicWoodcutter extends GenericJob {
 
 	@Override
 	public void simulate(Market market) {
-		float food = queryInventory("food");
-		float tools = queryInventory("tools");
+		double food = queryInventory("food");
+		double tools = queryInventory("tools");
 
 		boolean has_food = food >= 1;
 		boolean has_tools = tools >= 1;
